# openapi/fees/openapi3.yaml
openapi: 3.0.0
info:
  contact:
    name: Kita Knirpsenstadt
  description: API für Beitragsverwaltung der Kita Knirpsenstadt
  title: Kita Knirpsenstadt Fees API
  version: 1.0.0
paths:
...
  /parents:
    get:
      description: Get a paginated list of parents with optional search and sorting
      parameters:
        - description: Page number
          in: query
          name: page
          schema:
            type: integer
            default: 1
        - description: Items per page
          in: query
          name: perPage
          schema:
            type: integer
            default: 20
        - description: Search by name or email
          in: query
          name: search
          schema:
            type: string
        - description: Sort field (name, email)
          in: query
          name: sortBy
          schema:
            type: string
            default: name
        - description: Sort direction (asc, desc)
          in: query
          name: sortDir
          schema:
            type: string
            default: asc
      responses:
        "200":
          description: Paginated list of parents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParentList"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
        - BearerAuth: []
      summary: List all parents
      tags:
        - Parents
    post:
      description: Register a new parent in the system
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateParentRequest"
        description: Parent data
        required: true
      responses:
        "201":
          description: Parent created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Parent"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
        - BearerAuth: []
      summary: Create new parent
      tags:
        - Parents
  "/parents/{id}":
    delete:
      description: Remove a parent from the system
      parameters:
        - description: Parent ID (UUID)
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Parent deleted
        "400":
          description: Invalid parent ID
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Parent not found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
        - BearerAuth: []
      summary: Delete parent
      tags:
        - Parents
    get:
      description: Retrieve detailed information about a specific parent
      parameters:
        - description: Parent ID (UUID)
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Parent found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Parent"
        "400":
          description: Invalid parent ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Parent not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
        - BearerAuth: []
      summary: Get parent by ID
      tags:
        - Parents
    put:
      description: Update parent information
      parameters:
        - description: Parent ID (UUID)
          in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateParentRequest"
        description: Updated parent data
        required: true
      responses:
        "200":
          description: Parent updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Parent"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Parent not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
        - BearerAuth: []
      summary: Update parent
      tags:
        - Parents
  "/parents/{id}/member":
    delete:
      description: Remove the member association from a parent (does not delete the member)
      parameters:
        - description: Parent ID (UUID)
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Member unlinked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Parent"
        "400":
          description: Invalid parent ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Parent not found or not linked to a member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
        - BearerAuth: []
      summary: Unlink member from parent
      tags:
        - Parents
    post:
      description: Create a new club member from the parent's data and link them
      parameters:
        - description: Parent ID (UUID)
          in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMemberFromParentRequest"
        description: Membership start date (defaults to oldest child's entry date)
      responses:
        "201":
          description: Member created and linked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Parent"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Parent not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Parent already linked to a member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
        - BearerAuth: []
      summary: Create member for parent
      tags:
        - Parents
...
servers:
  - url: http://localhost:8081/api/fees/v1
  - url: https://localhost:8081/api/fees/v1
components:
...
  securitySchemes:
    BearerAuth:
      description: JWT Bearer token authentication
      in: header
      name: Authorization
      type: apiKey
  schemas:
...
    CreateMemberFromParentRequest:
      description: Request body for creating a member from a parent
      properties:
        membershipStart:
          example: 2023-08-01
          type: string
      required:
        - membershipStart
      type: object
...
    CreateParentRequest:
      description: Request body for creating a new parent
      properties:
        annualHouseholdIncome:
          example: 65000
          type: number
        birthDate:
          example: 1985-03-15
          type: string
        city:
          example: Potsdam
          type: string
        email:
          example: thomas.mueller@example.com
          type: string
        firstName:
          example: Thomas
          type: string
        incomeStatus:
          enum:
            - PROVIDED
            - MAX_ACCEPTED
            - PENDING
            - NOT_REQUIRED
            - HISTORIC
            - FOSTER_FAMILY
          example: PROVIDED
          type: string
        lastName:
          example: Müller
          type: string
        phone:
          example: +49 331 12345
          type: string
        postalCode:
          example: "14467"
          type: string
        street:
          example: Hauptstraße
          type: string
        streetNo:
          example: "42"
          type: string
      required:
        - firstName
        - lastName
      type: object
...
    ErrorResponse:
      description: Standard error response format
      properties:
        code:
          example: 400
          type: integer
        error:
          example: Bad Request
          type: string
        message:
          example: invalid request body
          type: string
      required:
        - code
        - error
        - message
      type: object
...
    Parent:
      description: Parent information with relationships
      properties:
        annualHouseholdIncome:
          example: 65000
          type: number
        birthDate:
          example: 1985-03-15
          type: string
        children: {}
        city:
          example: Potsdam
          type: string
        createdAt:
          example: 2023-01-15T10:00:00Z
          type: string
        email:
          example: thomas.mueller@example.com
          type: string
        firstName:
          example: Thomas
          type: string
        household: {}
        householdId:
          example: 550e8400-e29b-41d4-a716-446655440001
          type: string
        id:
          example: 550e8400-e29b-41d4-a716-446655440000
          type: string
        incomeStatus:
          enum:
            - PROVIDED
            - MAX_ACCEPTED
            - PENDING
            - NOT_REQUIRED
            - HISTORIC
            - FOSTER_FAMILY
          example: PROVIDED
          type: string
        lastName:
          example: Müller
          type: string
        member: {}
        memberId:
          example: 550e8400-e29b-41d4-a716-446655440002
          type: string
        phone:
          example: +49 331 12345
          type: string
        postalCode:
          example: "14467"
          type: string
        street:
          example: Hauptstraße
          type: string
        streetNo:
          example: "42"
          type: string
        updatedAt:
          example: 2023-01-15T10:00:00Z
          type: string
      required:
        - createdAt
        - firstName
        - id
        - lastName
        - updatedAt
      type: object
    ParentList:
      description: Paginated list of parents
      properties:
        data:
          items:
            $ref: "#/components/schemas/Parent"
          type: array
        page:
          example: 1
          type: integer
        perPage:
          example: 20
          type: integer
        total:
          example: 50
          type: integer
        totalPages:
          example: 3
          type: integer
      required:
        - data
        - page
        - perPage
        - total
        - totalPages
      type: object
...
    UpdateParentRequest:
      description: Request body for updating a parent
      properties:
        annualHouseholdIncome:
          example: 65000
          type: number
        birthDate:
          example: 1985-03-15
          type: string
        city:
          example: Potsdam
          type: string
        email:
          example: thomas.mueller@example.com
          type: string
        firstName:
          example: Thomas
          type: string
        incomeStatus:
          enum:
            - PROVIDED
            - MAX_ACCEPTED
            - PENDING
            - NOT_REQUIRED
            - HISTORIC
            - FOSTER_FAMILY
          example: PROVIDED
          type: string
        lastName:
          example: Müller
          type: string
        phone:
          example: +49 331 12345
          type: string
        postalCode:
          example: "14467"
          type: string
        street:
          example: Hauptstraße
          type: string
        streetNo:
          example: "42"
          type: string
      type: object
...